+++ src/SharpDbg.Infrastructure/Debugger/ManagedDebugger_TransientEvaluationModules.cs
@@ -1,0 +1,16 @@
+namespace SharpDbg.Infrastructure.Debugger;
+
+public partial class ManagedDebugger
+{
+	private readonly HashSet<Guid> _transientEvaluationModules = new();
+
+	internal void RegisterTransientEvaluationModule(Guid moduleVersionId)
+	{
+		lock (_transientEvaluationModules) _transientEvaluationModules.Add(moduleVersionId);
+	}
+
+	private bool TryConsumeTransientEvaluationModule(Guid moduleVersionId)
+	{
+		lock (_transientEvaluationModules) return _transientEvaluationModules.Remove(moduleVersionId);
+	}
+}
